refactor: remove repository once
